+++ spirit-log/src/background.rs
@@ -1,5 +1,10 @@
 // The mutex_atomic here is a false positive. We use the mutex because of the condvar.
-#![allow(unknown_lints, clippy::unknown_clippy_lints, clippy::mutex_atomic)]
+#![allow(
+    unknown_lints,
+    renamed_and_removed_lints,
+    clippy::unknown_clippy_lints,
+    clippy::mutex_atomic
+)]
 //! Support for logging in the background.
 //!
 //! The [`AsyncLogger`] can wrap a logger and do the logging in a separate thread. Note that to not
@@ -228,6 +233,15 @@ pub enum OverflowMode {
     },
 }
 
+impl OverflowMode {
+    fn count_lost(self) -> bool {
+        matches!(
+            self,
+            OverflowMode::DropMsg | OverflowMode::AdaptiveDrop { .. }
+        )
+    }
+}
+
 /// A logger that postpones the logging into a background thread.
 ///
 /// Note that to not lose messages, the logger need to be flushed.
@@ -332,15 +346,14 @@ impl Log for AsyncLogger {
             };
             if self.mode == OverflowMode::Block {
                 self.ch.send(i).expect("Logging thread disappeared");
-            } else if let Err(TrySendError::Full(_)) = self.ch.try_send(i) {
-                match self.mode {
-                    OverflowMode::DropMsg | OverflowMode::AdaptiveDrop { .. } => {
+            } else {
+                match self.ch.try_send(i) {
+                    Err(TrySendError::Full(_)) if self.mode.count_lost() => {
                         self.shared.lost_msgs.fetch_add(1, Ordering::Relaxed);
                     }
-                    _ => (),
+                    Err(TrySendError::Full(_)) | Ok(()) => (),
+                    _ => panic!("Logging thread disappeared"),
                 }
-            } else {
-                panic!("Logging thread disappeared");
             }
         }
     }