+++ spirit-log/src/lib.rs
@@ -3,6 +3,7 @@
 #![warn(missing_docs)]
 #![allow(
     unknown_lints,
+    renamed_and_removed_lints,
     clippy::unknown_clippy_lints,
     clippy::needless_doctest_main
 )]
@@ -912,6 +913,8 @@ impl<O, C> Installer<Dispatch, O, C> for LogInstaller {
         install(logger);
     }
     fn init<B: Extensible<Ok = B>>(&mut self, builder: B, _name: &str) -> Result<B, AnyError> {
+        #[cfg(feature = "background")]
+        let builder = builder.with_singleton(FlushGuard);
         builder.with(Cfg::init_extension())
     }
 }
@@ -922,6 +925,8 @@ impl<O, C> Installer<(LevelFilter, Box<dyn Log>), O, C> for LogInstaller {
         install_parts(level, logger);
     }
     fn init<B: Extensible<Ok = B>>(&mut self, builder: B, _name: &str) -> Result<B, AnyError> {
+        #[cfg(feature = "background")]
+        let builder = builder.with_singleton(FlushGuard);
         builder.with(Cfg::init_extension())
     }
 }